TO DO consertar listas

+++ locacao/locacao_informacoes.dart
@@ -2,7 +2,7 @@ class LocacaoInformacoes{
   int? codigoLocacao;
   var dataSaidaLocacao;
   var dataEntregaLocacao;
-  int? diasLocadosLocacao;
+  double? diasLocadosLocacao;
   int? idVeiculoLocacao;
   int? idClienteLocacao;
   int? idFuncionarioLocacao;
